requirements.txt
gtts==2.5.4
plotly==6.5.2
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.54.0
